YouTube Static Project
YouTube Static Project
Welcome to the YouTube Static Project, a representation of a YouTube webpage using only HTML and CSS. This project focuses on showcasing the core elements and styling techniques necessary to replicate the appearance of a standard YouTube interface. Dive into the HTML and CSS code to gain insights into how this static representation was achieved.
Key Features:
HTML Structure: Demonstrates the use of HTML to organize the content and layout of a YouTube-like webpage, including headers, sections, and lists.

CSS Styling: Utilizes CSS to apply styles and achieve the visual design elements such as colors, fonts, and layout positioning to mimic the YouTube interface.

Technologies Used:
HTML
CSS
How to Use:
Clone this repository to your local machine.
Open the index.html file in your preferred web browser to view the static YouTube webpage.
Explore the HTML and CSS files to understand the structure and styling employed.
Contributing:
If you would like to contribute or provide feedback, feel free to fork the repository and create a pull request.